vopr: fix false positive when getting to recovering_head after version upgrade
We have special code to disable storage faults when restarting a replica
after a crash in the vopr. We also have similar code when restarting
replica during upgrades but it doesn't quiet handle all special cases
(in particular, it doesn't repair the WAL). Extract this special code
into a special function to make sure that more of the logic is shared!

Seed: ./zig/zig build -Drelease vopr -- 11029999072009206202