LSM: Optionally index zero values
As of https://github.com/tigerbeetle/tigerbeetle/pull/2220, zero-amount transfers are allowed.
Since we have an index on `Transfer.amount`, we need to ensure zero-amount transfers are indexed too!

Note that we don't support querying by amount yet, so this isn't actually tested.

This commit should be merged before `0.16.0` is released.