Derived index function returns an optional.
This change is motivated by the need to index flags that can return either `void` (exists) or `null` (skip the index).

Derived functions used to return an Optional in the past, but this was changed to treat zero as optional instead. Now that we explicitly handle optional indexes, reverting the signature makes the intent even clearer.