Change init Signature
This is a different (and simpler) approach compared to PR #2242, avoiding copying and preventing stack allocation of local objects by changing the init signature to take the object's pointer.

While the overall object size remains the same (e.g., Forest is still ~608KiB), this solution is even more flexible, allowing the caller to choose whether to allocate it on the stack or the heap. The initialization is guaranteed to preserve stack size, avoiding both copying and moving data around.