vsr: prefer wal repair over state sync
New state sync protocol regressed the behavior where the replica would
try to repair WAL switching to state sync, and this commit puts the old
behavior back in.

Good thing that we have a test here! Sadly, the test was silently
regressed as it didn't actually check that the state sync didn't happen.
Mark the relevant logs to pin down the behavior!

As for the actual fix, it centers on the `repair_stuck` function.
Before, it was a stateless function. My original thinking was that we
will only get an SV message in response to our own RSV, and, given that
we have a repair_sync_timeout when requesting an SV, there should be
natural debouncing here.

What I didn't consider are the two other paths to send an RSV:

* when receiving a message from a newer view
* in `.repair()` when our head is < commit_max

So what actually happens is that a lagging replcia joining the cluster
would pretty much spam the primary with RSVs anyway, which in turn,
would immediately flip the laggard into the state sync.

So, in this PR `repair_stuck` tracks a bit of state. Namely, once every
repair_sync_timeout we note our commit_min, and note whether it is
_different_ from the one we had before. commit_min being the same across
two timeouts becomes another ingredient to decide if the commit is
stuck.

This state is tracked in `sync_wal_repair_progress` which is
intentionally never reset and only ever updated by the timeout.