build: vendor llvm-objcopy
We now depend on llvm-objcopy to make our multiversion binaries. Long
term, we should probably upstream the capabilities we need to `zig
objcopy` but for now "vendor" a specific copy of llvm-objcopy we use.

Vendoring is useful primarily because there isn't an easy way to install
the tool in a cross-platform way! As we want to make multiversion a
normal build, it is important that the build remains seemless and that
the user doesn't have to manually install anything.

The solution is to upload specific versions of llvm-objectopy as a
release on <https://github.com/tigerbeetle/dependencies> and use that.

Sadly, you have to trust me that the binaries I uploaded there are good,
but the mac and windows versions are straight from LLVM upstream
release, and the build of statically-linked linux version seems to be
reproduciable.

Assuming the specific binaries are fine, any further risks are locked
down by veryfying the checksum upon the download.