vsr: remove obsolete sync_view field
This was a part of the old state sync protocol, where the role of the
field was to communicate in which view the sync target checkpoint was
committed.

Today, this information is tracked explicitly in an SV message which
allows us to atomically:

* advance our view
* install WAL consistent with the view
* install the checkpoint consistent with the WAL

So, we can remove `sync_view`. Sadly, that is a part of our superblock,
so we still need to have _something_ on disk there. My plan is to mark
this field `reserved` and assert that it is zeroed, but, before doing
that, lets explicitly zero it out. With all probability, it is already
zeroed everywhere in practice, but it is theoretically possible that a
replica went through several upgrades while not finishing state sync, so
let's lean on the safe side here.