docs: used # for document titles and ## for section headings
There are two schools of thought regarding largest (`# War and Peace`)
headings in markdown. One school says that they are _section_ titles, so
you could have multiple #'s in the document. But another option is to
say that a single # signifies document _title_, and there should be only
one in a document.

We use markdown to create HTML, so # turns into h1. MDN recommends that
there's only a single h1 in a page:

<https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements#avoid_using_multiple_h1_elements_on_one_page>

For this reason,  we follow the second convention.