build: remove default development build mode
The semantics of default_development are not well-defined --- is it for
TigerBeetle developers, or for developers _on_ TigerBeetle?

In any case, this is a philosophical question, as our development is
equal to production except for `verify` flag, _and_ we actually run with
`verify=true` in production still!

So let's get rid of default_development, and formally say that we do
default verify to true.