vsr: refactor commit_dispatch loop
Rewrite commit_dispatch to more closely resemble straight line  code.

 # Motivation

While I like the general idea of commit_dispatch, it feels like our
current implementation isn't quite as obvious as it needs to be:

* Each stage is responsible for calling the next stage (you pass an
  argument to .commit_dispatch), which makes it hard to know in which
  order things actually happen. They sort-of happen in the order of the
  switch prongs in the dispatch, but it's not immediately clear if we
  are not being clever and skipping some of the stages sometimes.
* I always have trouble finding the place where we _actually_ execute
  things. The entry point is `setup_client_reply_callback`, which is the
  least likely springboard for the actual execution!
* There's some confusion about "sometimes asynchronous" stages, up to
  the point where we get unexpected-reentracy _despite_ having an
  explicit comment in the source code about _not_ doing that.

 # Implementation

The new approach is an attempt to do the most direct thing possible. If
programming were actually easy, we'd write our commit path as, roughly

```
loop {
  prefetch()
  execute()
  compact()
}
```

The problem is that programming is, in fact, not easy, and the principle
difficulty here is that most of parts of the loop are asynchronous. Were
we using a language with async-await, we would have written:

```
loop {
  prefetch().await
  execute().await
  compact().await
}
```

We don't have async-await, but we can express the same logic manually.
`prefetch` being async means that `prefetch` might not actually complete
immediately, and instead just schedule some background work. Which means
that we need to "pause" our commit work and return to it later.

Pausing is easily achieved by early returning:

```
loop {
  if (prefetch() == .pending) return
  if (execute() == .pending) return
  if (compact() == .pending) return
}
```

Resumption is trickier, as we should be able to restart the loop from
the middle. If we were writing in C, we could have done something like
this:

```
switch(state) {
  loop {
case prefetch:
    if (prefetch() == .pending) return
case execute:
    if (execute() == .pending) return
case compact:
    if (compact() == .pending) return
  }
}
```

Luckily, Zig doesn't allow this kind of convoluted code (though I remain
intrigued by the possibilities opened by the recently implemented
https://github.com/ziglang/zig/issues/8220).

But we can simulate it easily enough by guarding each prong of the
switch with an if:

```
loop {
  if (state == .init) {
    state == .prefetch
    if (prefetch() == .pending) return
  }
  if (state == .prefetch) {
    state == .execute
    if (execute() == .pending) return
  }
  if (state == .execute) {
    state == .compact
    if (compact() == .pending) return
  }
}
```

Which is exactly what are we doing here!

Specifically:

* `commit_dispatch` is the main commit loop. It is an asynchronous loop
  committing everything we can commit.
* `commit_dispatch_enter` enters the loop from the outside.
* if the loop schedules an asynchronous operation, it's callback should
  invoke `commit_dispatch_resume` without arguments, which will pick up
  the loop where we've left.
* All the "parts" of the `commit_dispatch` loop are right after it in
  the source code order and are named `commit_stage`. All callbacks are
  named `commit_stage_callback`

One interesting side-effect here is that the new structure will be able
to accommodate our "optimistic prefetch" idea, where we _assume_ that
everything is in cache and start commit _without_ prefetch, bailing out
if we have a missing entry. The new loop works even if the entire commit
path turns out to be synchronous!