[StateMachine] Transient failure This introduces the new concept of "transient failure," which are events that fail due to the system's status rather than the request data. The list of transient errors is as follows:
create_accounts: none
create_transfers: `debit_account_not_found`, `credit_account_not_found`, `pending_transfer_not_found`, `exceeds_credits`, `exceeds_debits`, `debit_account_already_closed, and `credit_account_already_closed`.

With this, we introduce the invariant: "Once a transfer fails with a transient error code, it must always fail with `id_already_failed` when retried with the same ID." This ensures that a user will never encounter a "transient error code" twice for the same ID, providing the application with stronger guarantees that the outcome of a transfer will remain consistent across retries.

Additionally, this changes the precedence of the idempotency checks (i.e., all `exists_*` checks) to the beginning, right after `id_must_not_be_zero` and `id_must_not_be_int_max`.

The objectless `id` is stored in the `IdTree` with a zeroed `timestamp`.

The Groove is not aware of "transient failures" but instead it has a configuration that enables "orphaned IDs", which are IDs stored without a corresponding object.
We are using the ObjectCache (with an empty object) to cache such IDs, which can be beneficial for pathological workloads that retry the same failed event in a loop.