test: run concurrent workloads with separate ledgers
Restructures the systest workload (in Java) to run multiple workloads in
separate threads, isolated by using unique ledgers. This let's us verify
the thread safety claim of the Java client, which is shared across the
threads.

The verified properties after querying known accounts are similar to before, because we already
grouped accounts by ledger and checked consistency for each group.