Multiversioning: Read the entire binary
The `"in-place upgrade"` integration test went from passing consistently to failing _most_ (though not all) of the time.

In the upgrade integration test, after we replace the binary, the running replica detects that the binary changed.
Then it reads the new binary.

When the whole binary fits in a single read, the test passes.
When the whole binary doesn't fit in a single read, it fails to parse (due to `InvalidELF`).

(I'm guessing that this started failing as the binary has grown larger (as more versions are bundled each week).)

Fix: Use the size from `statx()`, and keep reading from increasing offsets until the entire binary is read.