state_machine: fix cross-release bug, tests, rename FormerAccountFilter
After adding new fields to the `AccountFilter` in 0.16.2, a
`FormerAccountFilter` was added for compatibility with older clients.

Its usage was determined by `account_filter_from_input()`, which
switches appropriately depending on the filter type, but this fn
wasn't called in the execute path for `execute_get_account_balances`.

Instead, it always tried to interprete the filter as the new type,
leading to a panic when executing `get_account_balances` on a replica
running 0.16.2 or newer from an older client.

This commit fixes that bug, adds a few tests to catch it, and adds the
ability to `setup_client_release` in state_machine unit tests for
testing cross-release things like this. All it does currently is set
the client_release passed in to things like `input_valid` and
`prefetch`.

While we're here, `FormerAccountFilter` -> `AccountFilterFormer`.