[Clients] Don't batch unclosed linked chains
If the application submits an unclosed linked chain, it can inadvertently make the elements of the next batch part of the same linked chain. To prevent this, do not allow unclosed linked chains to be batched together with other logical batches in the same request. Instead, allow the StateMachine to handle the validation with `.linked_event_chain_open`.

Note: This correction will be unnecessary once protocol batching is implemented.