test: write results from main thread in Zig driver
Instead of writing to stdout from the tb_client callback directly, we
use a mutex and a condition to await the results in the main thread,
through a shared context.

One reason for doing this is that we can reinitialize the client in
a controlled way. Previously, a callback might happen after
reinitializing the client.

This is based on the same mechanism as used in
`src/clients/c/samples/main.zig`.