test: stop/continue replicas in vortex supervisor
In order to try how replicas can recover from being not only crashed
(kill and restart in the case of Vortex), but also from being stopped
and then resumed, this PR adds new faults.

It adds a new `stopped` state to `LoggedProcess` and `Replica`, and
also removes the already unnecessary "completed" state entirely.