Multiversion: Simplify upgrade with dev flags
## Overview

When `--development` or `--experimental` is set, upgrading is more complicated than usual:

1. Stop the replica (old binary).
2. Start the replica (new binary) without the offending flag.
3. Wait for the upgrade to complete.
4. Stop the replica (new binary).
5. Start the replica (new binary) with the extra flags.

In this PR, simplify it to:

1. Stop the replica (old binary).
2. Start the replica (new binary) with the extra flags.

i.e. when `--development`/`--experimental` are set, don't monitor the binary for changes, but _do_ support upgrades.

Also in this commit, clarify some of the `log.info` messages.

## Details

Note that the previous logic:

```zig
if (args.development) {
    log.info("multiversioning: disabled due to --development.", .{});
    break :blk null;
}
if (args.experimental) {
    log.info("multiversioning: disabled due to --experimental.", .{});
    break :blk null;
}
```

... still exists in the bundled versions, so upgrades with `--development`/`--experiment` will only work as those bundled versions get phased out, replaced by binaries with the new logic.

To test this locally, I first built an "old" version with the new logic, then bundled that "old" version with the "new" binary, and tested an upgrade.
(See the change in `build.zig` in the prior commit, which simplifies this process).