client: expose reply timestamp
Fill-in the API gap, allowing the clients to unambiguously order all
replies.

Before this PR, if two different clients issue a `.lookup_accounts`
call, it would be impossible, using only the public API, to order them
relative to each other.

That is, _one_ of the requests completed first, and it can be determined
from the timestamp of the VSR reply message, but that timestamp is not
exposed anywhere.

This currently isn't threaded all the way up to language clients, to
keep the scope smaller.