io: fix race condition in fs_supports_direct_io
This fixes (at least one issue) flaky upgrade test. There, several
replicas are being run in the same directory, and they all try to create
this `fs_supports_direct_io` probe file, which mostly succeeds, but
sometimes can fail!

Add a random cookie to prevent this.

And, as this is the 1274278th time I need to print a number to a
stack-allocated buffer, also add an `stdx.array_print` utility to check
buffer's size automatically.