style: emphasize human reasoning
Recently, VOPR became to good, and it becomes _really_ tempting to
switch to an _empirical_ mode of coding: write some code and let the
VOPR figure out whether it is correct or not.

This is suboptimal, for two reasons.

_Firstly_, VOPR, as any testing methodology, is one-sided --- if it
finds a bug, there's definitely a bug, but silence of the VOPR doesn't
guarantee total absence of bugs! Something might be _really_ unlikely,
or completely uncovered by the VOPR.

_Secondly_, safety comes in layers and cross checks. Just formal or
informal reasoning is not enough. And just simulation is not enough.
Simulation AND reasoning is still not enough, but is _much_ better,
because a bug now has to slip trough two cracks in two different
surfaces! Minas Tirith is a seventh-fold walled fortress for a reason!

So instead of using VOPR to run experiments, like a physicist would do,
we should be more like a mathematician --- derive correctness of code
from the first principles, and use VOPR to debug our understanding
instead of debugging our code.