Message: rename .body() to .body_used(), and use it for state_machine.commit
This matches the convention inside the LSM, and makes it more clear at a
glance that you're getting the portion of the body bounded by the
header's size.

This was prompted by seeing that replica.zig calls the state machine with
a manual `prepare.buffer[@sizeOf(Header)..prepare.header.size]` and wondering
why that didn't use the original `.body()`. Turns out, there doesn't seem
to be a reason, so change that to use `.body_used()` too.