Journal: Fix assertion after misdirected read/write
Previously we asserted `slot_with_op_and_checksum(op, checksum) == null` in these cases, because it would be `null` in the case of "prepare was rewritten since the read began".

But due to a misdirected write, the header may be present, and the message in the WAL is just wrong.

(Both of these assert failures are findable by the VOPR with misdirected writes enabled.)

Seed: ./zig/zig build -Drelease vopr -- 8358894577611284529