log: change comptime log to .debug and add runtime filtering
Previously, .debug log messages were filtered out at comptime. While nice
from a binary size and performance (more on that later) point of view,
this meant that a user couldn't specify `--log-debug` or similar and would
need a new binary to get debug logging.

Make it so that we do runtime log filtering instead. The performance
penalty for this was negligible in testing; both microbenchmarks and the
actual benchmark.

The one downside is binary bloat: this increases the size of a release
build by ~300KiB.

To use, specify `--log-debug --experimental` on the CLI when starting
a replica.

It's currently only supported for `start`. Scaffolding for clients is there
too, with filtering of the messages before they cross FFI and requiring
an additional flag to be set. Actually setting this flag from any client
libraries is still unimplemented - but this makes it no harder than
before to change client logging (edit code & recompile).