message_bus: fix asymmetrical TCP disconnection handling
When a client reconnects, it uses the same client id, and this is
hanlded internally by `set_and_verify_peer()`: the old connection
in the bus's connection map is replaced by the new connection and
the old connection has its `.terminate()` called.

`.terminate()` shuts down the socket using `posix.shutdown` and
eventually calls `maybe_close()`, either directly - which often
does nothing since there's an outstanding recv - or by setting
the connection state to `.terminating` which results in
`maybe_close()` getting called by the `on_recv` & co handlers
when the completion comes in and the socket can be safely closed.

`maybe_close()` eventually calls `io.close()`, and in the `on_close`
callback, removes the connection from the client map.

The bug arises because this connection that's being removed is being
done by client id, and it's not necessarily the same old physical
connection, in fact, given what happened above, and that `on_close`
is always invoked asynchronously, it's the _new_ connection object!

When that gets removed, nothing bad happens yet, until the real
connection closes, and the assert is triggered. Interestingly, this
case was already handled for a replica peer, so use that same logic
to ensure we only remove the connection if it is the same physical
one.

(The problem only manifests with asymmetrical TCP disconnections -
where the client knows its disconnected, so it reconnects, but the
replica hasn't yet received the event - so it has a connection
object for the existing client id still.)