tidy: ban `!comptime ...`
Sometimes, you need to reach for an explicit `comptime` in an if condition:
```
if (comptime foo) {
    std.log.info("foo!", .{});
}
```

and sometimes you need to negate that:

```
if (!comptime foo) {
    std.log.info("not foo!", .{});
}
```

all good so far. The problem is that this comptime expression spans across
the entire condition, so:

```
if (!comptime foo and bar) {
    std.log.info("not foo, bar!", .{});
}
```

is really `!comptime (foo and bar)` and not `!(comptime (foo)) and bar.