Journal: Don't `cut_view_range` for redundant headers
# Bug

In `Journal.recover_slots()`, we set the decision to `TruncateViewRange` if _either_ the prepare header or redundant header has an out-of-range view:

```zig
for ([_][]align(constants.sector_size) Header.Prepare{
    journal.headers_redundant,
    journal.headers,
}, 0..) |headers, list| {
    for (headers, 0..) |*header_untrusted, index| {
        const slot = Slot{ .index = index };
        if (header_ok(replica.cluster, slot, header_untrusted)) |header| {
            var view_range = view_change_headers.view_for_op(header.op, log_view);
            assert(view_range.max <= log_view);

            if (header.operation != .reserved and !view_range.contains(header.view)) {
                cases[index] = &case_cut_view_range;
            }
        }
    }
}
```

This can cause problems -- suppose:

1. Write prepare.
2. Write redundant header. (Misdirected write.)
3. Send `prepare_ok`.
4. Crash, restart, recover WAL...

...then we set the decision to `TruncateViewRange` (i.e. `cut_view_range`), which sets the recovered header to `reserved` and clears both `dirty` and `faulty`.

In the VOPR, this is caught by `state_checker.zig`:

```zig
assert(replica.view > head_max.view or
    (replica.view == head_max.view and replica.op >= head_max.op));
```

That is, a replica may truncate an op that it already acked due to a single misdirected write!

# Fix

The simple solution here is just don't apply `TruncateViewRange` based on the redundant headers.

But taking a step back, I think that the `TruncateViewRange` "filter" is being applied too early in recovery.
It is applied before the recovery decisions are computed, but the recovery decisions interpret any `reserved` headers as _actual_ reserved headers, not as " of a header we didn't mean to write".

Instead, move the view-range check after decisions are computed (but still before `recover_fix()` is applied ­ we don't want to "fix" headers that we are just going to discard).
This means that there is no question of testing both the redundant and prepare headers -- after the decisions are computed (`recover_slot()`) they are unified in `headers`.

So in the scenario described earlier, we would end up in case `@L` or `@M`.

(Interesting seeds: 7056965896392806540 and 4595027732159043625.)