clock: change log severity levels
Upgrade `no agreement on cluster time` to an error: it means requests can't
be processed _right now_ and it's as good of a reason as any to page
someone.

At the same time, change the synchronized message to an info: you'll want
to be able to see that things have returned to normal.