time: explain how asymmetric delay computation works
I had a completely wrong mental model for what we are doing here. I
_assumed_ we were after _stationary_ asymmetries, where sending packets
there is always slower than getting packets back.

But it looks like this kind of asymmetry isn't actually possible to
repair!

Instead, what the code does is fixing _temporary_ delays, when, in an
otherwise normal ping-pong stream, something temporary goes wrong and a
single message gets delayed. _This_ case we can detect by observing RTT
going way up!

To be honest, I am unconvinced at the moment that this is a useful thing
to do: to detect asymmetric delay, we rely on a previous sample with
lower RTT. But if we use that sample to estimate delay anyway, we could
also use it for its clock value? I'd expect network weather to be much
more volatile than time, so, if we judge the past sample to be fresh
enough to estimate network delay, it should be fresh enough to judge
time either! But I am not quite ready to remove that fence!