clock: don't let large RTT mask unsynchornized clocks
Although we do binary search over the `tolerance` variable, the _actual_
error margin we use for Marzullo is `tolerace + RTT`, as seen in
`window_tuples`:

    .offset = sample.clock_offset -
        @as(i64, @intCast(sample.one_way_delay + tolerance)),

That is, if RTT is large, it would _easier_ for Marzullo to find a
quorum of clocks.

So, if pings take 30 seconds to round-trip, we might end up with the
interval 50 (= (10 + 30/2) * 2) seconds wide, larger than our max offset
tolerance of 10 seconds.

Fix this by capping the synchronized interval to the offset tolerance.