VSR: Set upgrade quorum size to replica_count
Previously the upgrade quorum (i.e. the number of replicas which must agree to upgrade before the cluster will begin to upgrade) was `replica_count - 1`.

This decision predates the idea of "hotfix releases" – builds which include a bugfix and nothing else, which are tagged with the same release number as their predecessor, and so don't require an upgrade quorum/commit/checkpoint.

Since upgrade a partial cluster is pretty much never what the operator wants, set the quorum size to `replica_count`.

As it happens, this is already what we had documented in `upgrading.md`:

> TigerBeetle will then coordinate the actual upgrade when all replicas are ready and have the latest version available.