docs: Add note to docker server startup process
In #2509, @ashley-nygaard had an issue connecting a local TB client (running on macos) to a TB replica running in Docker.
After some troubleshooting, we determined that the solution was to explicitly set the `--address` parameter to the IP address of the Docker container.
Updating the documentation with this note to make future developers aware.