cfo: ignore OOM failures
A single fuzzer can use more than 16 gigs of RAM under some
swarm-testing configurations, but the average is expected to be much
smaller than that.

With many fuzzers running concurrently, the total CFO memory consumption
can be unreasonably high. While we can pre-allocate RAM for the worst
case, this will significantly reduce resource utilization.

For this reason, we will explicitly "catch" OOM and ignore OOMing seeds.
OOM corresponds to getting SIGKILLed, and, as far as I am aware, no
internal error can manifest itself as a sigkill, so this is safe.

An alternative solution here would be to disable overcommon, and to
carefully handle allocation errors in the fuzzers. I _think_ that this
will lead to significantly more complicated code of no real benefit.