LSM/Scan: Fix `ScanMerge.move_next()` assert
This assert is hit by nested alternating intersection/unions.

Suppose there is a query `(index_0 AND (index_1 OR (index_2 AND index_3)))` with the listed timestamps in each index:

    zig_zag_merge₁:     [              12, 13     ]
      tree₀:            [          11, 12, 13     ]
      k_way_merge₁:     [ 1, 2, 3,     12, 13, 14 ]
        tree₁:          [ 1, 2, 3,                ]
        zig_zag_merge₂: [              12, 13, 14 ]
          tree₂:        [ 1, 2, 3,     12, 13, 14 ]
          tree₃:        [              12, 13, 14 ]

1. While peeking the first key from `zig_zag_merge₁`, we peek `11` from `tree₀` and `1` from `k_way_merge₁`.
2. So we probe `k_way_merge₁` with `11`. `k_way_merge₁` relays `11` to its streams (`tree₁` + `zig_zag_merge₂`).
3. Probing `zig_zag_merge₂` with `11` trips the assert, because `tree₃` has already produced a higher key (`11 < 12`).

(Note that we can't simply gate this assert on `merge == .merge_intersection`, since the inner intersection still fails.)

This assert wasn't hit until now because probing in the scan fuzzer wasn't skipping over any keys.
(It can't be hit by the VOPR or the actual database since we don't use this structure of query in the state machine.)
The updated scan fuzzer hits this bug promptly.