misc: use --save-exact for node and better log on eviction
Without `--save-exact`, node will record the package as eg `^0.16.16`
which will bump patch versions. This is incompatible with how we handle
versions between clients and replicas.

Additionally, note that down in the release notes, and include Python
there which was missing.

Lastly, when logging the eviction, include the client release and hint
on what to do, if it was due to mismatched versions.