vsr: fix a crash when the root prepare gets corrupted
On this seed, a replica is on the initial checkpoint, and everything it
prepared (3 prepares in total) got corrupted, including the root
prepare.

Normally in this case the replica uses op_checkpoint for op_head, but,
as `self.journal.op_maximum() < self.op_checkpoint()` is `0 < 0` is
false, we skip the relevant code block, and end up in a situation when
`op_head == 0 and journal.header_with_op(0) == null`, tripping an
assert.

The root cause is that `op_maximum` should really be returning `null`
here, rather than `0`, to disambiguate the case where `0` is indeed the
maximum op, and the case where we don't have a maximum op at all.

I considered changing its return type, but then I'll also have to change
the return type of `op_maximum_untrusted`, and add a bunch of special
cases for no real benefit.

So, instead I just special-case a single if.

I don't understand why this was uncovered by pausing (parent commit) ---
I _think_ this should be findable on main as well, but perhaps with
lower probability.

Seed: ./zig/zig build vopr -Drelease -- --lite 657199375419925836