vsr: don't crash if a round of view changes happened while repairing pipeline
This fixes availability issue in VSR, where the prospective primary
could crash during view change when a journal read initiated in the
previous view change finishes. This is unlikely to be hit in practice,
but was discovered by VOPR.

 # Scenario

The prospective primary needs to fully repair before starting a new
view. The final part of the repair is the pipeline --- the prospective
primary needs to make sure that all potentially committed ops are in
memory. Specifically, they are to be stored in PipelineCache, which is
to be flipped over to PipelineQueue proper when the view is started.

The prospective primary _gets_ the prepares using the normal repair
protocol, but it could be the case that a prepare only exists on disk,
and not in memory.

Fetching prepares from disk is exactly the job of
`primary_repair_pipeline` function.

Reading prepares is asynchronous, so it could be the case that, by the
time the read finishes, it is no longer relevant. In the simple case,
the view is advanced by one and the replica is no longer future primary.

The more complex case is when a view is advanced by a multiple of
`replica_count`, so the replica is a prospective primary _again_.

In general, we can't say if the read we are dealing with was initiated
for _this_ view change, but we don't need to. It's fine if the older
reads _help_ a future view change, as long as any reads which are not
helpful are ignored.

This latter case was buggy, we could have proceeded with
repair_pipeline_callback even though we wouldn't have called
repair_pipeline from repair yet.

This commits adds the missing condition of commit_min == commit_max

 # Why wasn't this discovered before

This is a pre-existing bug. I _think_ it was brought to light through
pauses. CEO caught it where cfo missed :-)

Seed: ./zig/zig build vopr -- --lite 9406017288286641868