main: switch to GPA as base allocator
The arena allocator likes to overallocate quite generously. This isn't
normally a problem, but after enabling mlockall() in #2478, that unused
but mmap'd memory is now forcefully paged in and contributes to our
real RSS.

With default grid cache, this takes memory usage from 2.9GiB to 4.7GiB!
With a large grid cache (24GiB) it takes a whole extra 3GiB of memory.

Switch to using Zig's general purpose allocator, which overallocates
much less. This is still a bit of a workaround - but one that seems to
work rather well. A nice side-effect is that it makes it much easier
to catch leaks and use-after-frees, like the little fixes you'll see
included in this commit.

In terms of usage, both with the default grid cache size and a 24GiB
grid cache, the RSS is now identical with and without mlockall.

The performance difference isn't noticable when starting with the default
grid cache, and grows to around ~30% slower for 24GiB grid cache vs
the arena allocator.

This can be solved in future by optimizing how the grid cache allocates
blocks...