[StateMachine] Make alignment checks operation-specific. Accept non-aligned memory and check the alignment required by the specific state machine operation.
Previously, we enforced the state machine signatures to receive aligned slices as `[]align(16) u8` and `*align(16) [message_body_size_max]u8`.

This approach would cause friction with custom state machines by requiring all events/results to be aligned by a constant. It's also already inconsistent with results such as `CreateAccountsResult` and `CreateTransfersResult`, which are 8 bytes long. For example, when replying 3 results, we don't add any padding to keep it `align(16)`,  we would return 24 bytes instead of 32 bytes.

However, the main motivation for this change is protocol batching. When passing chunks of memory to the state machine for reading/writing, those chunks are aligned to the corresponding Event and Result, but not always aligned to 16 (especially when writing results to a buffer of `CreateAccountsResult` and `CreateTransfersResult`).