vsr: assert against ABA problem during commit
Reading prepare from the journal is asynchronous operation. Generally
speaking, there's no guarantee that, by te time the read finishes, the
prepare remains correct. E.g., you can start reading from slot=20  in
view=5, but finish in view=6, where that slot would have been occupied
by a different prepare with the same op number.

This actually cannot happen on the commit path, but for a very
non-trivial reason (general VSR correctness), so let's assert this via
`journal.has` which will compare the checksum of the prepare we have on
hand with the checksum we are supposed to be having.

Add the assert to the read callback and prefetch. We already assert
later this in execute.

Inspired by #2571, where a past read from a specific slot in journal was
interpreted as a recent one.