io: add retry for flock
On Linux, TigerBeetle takes out an `flock` over the data file when starting.
This turns out to lead to interesting races in CI or tests, when killing and
starting replicas in an automated fashion in quick succession.

`flock`s are cleaned up by the kernel when the file descriptor is closed,
but since that file descriptor is used by io_uring, it actually outlives
the process itself.

That means it's possible to kill a tigerbeetle replica, have it really
be dead, start another one on the same data file and get a
`another process holds the data file lock` error!

There's unfortunately no real satisfying way of fixing this nicely that
I could find; so I thought the simplest, a tiny backoff loop, might be
the best.

It is actually possible to mix in fcntl per-process locks, and check
both: if the fcntl lock doesn't exist, but the flock does, you can
then conditionally back off. But, this is a fair bit of extra
complexity.

50ms chosen as a backoff value because in practice I've seen it be
released within ~3ms relatively consistently.