VSR: Request extra headers during repair
When our journal has a break in the headers, don't repair the breaks serially.
Instead request as many headers as possible, to hopefully repair multiple breaks.

Suppose a backup's journal header suffix have a pattern like: `1 _ 3 _ 5 _ 7` (where `_` indicates a gap where we don't know the header). Instead of requiring three separate `request_headers`/`headers` round-trips, we can do it in one.

This alternating header/gap pattern is common in a cluster of three when the ring replication direction alternates and one of the backups is unavailable.