LSM: Defer compaction for small immutable tables
# Context

## Coalescing

[Greedy table coalescing](https://github.com/tigerbeetle/tigerbeetle/pull/2070) reduced space amplification of inserting objects with partially-full batches. (Especially for disjoint tables.)
It accomplished this by aggressively compacting small tables together until they aren't small anymore. (Even if the tables are only adjacent without actual overlap.)

Unfortunately, this extra compaction adds a lot of extra CPU/IO (especially write amplification) work to the cluster. The smaller the batches, the more times we rewrite the table.

## Immutable table compaction

Right now:

- We convert the mutable table into the immutable table after the last beat of each bar. (Aka the last beat of each second-half-bar.)
- But we don't actually start compacting the immutable table into level 0 until the first beat of each second-half-bar (odd → even).

For clarity:

```
[          bar 1         ][          bar 2         ]
[half-bar 1A][half-bar 1B][half-bar 2A][half-bar 2B]
```

The immutable table created at the end of *bar 1* (at the end of *half-bar 1B*) is flushed during *half-bar 2B*.
The immutable table isn't used (except for reads) during *half-bar 2A*.

# Optimization

At the end of each first-half-bar (e.g. *2A*), if the sum of:

- the number of values in the immutable table (constructed at the end of *1B*),
- the number of values in the mutable table so far (during *2A*), and
- the worst-case number of values that may be inserted into the mutable table during the last half-bar (*2B*)

...fits within a single table, then _don't compact the immutable table_.

Instead, at the end of the last half-bar (*2B*):

1. append the mutable table's values to the immutable table,
2. sort+dedup the immutable table, and
3. reset the mutable table.

### Caveats

- This optimization must be disabled during the bar immediately preceding each checkpoint trigger.
  - (Recovering from the checkpoint replays only the last bar of ops, and it must reconstruct the same immutable table.)
  - Because of this, LSM must now be aware of the checkpoint schedule. (Hence the forest/tree fuzzer modifications.)
- This optimization will need to be disabled when we are taking a persistent snapshot.
- Note that this doesn't provide a _guaranteed_ lower-bound table size.

# Performance

This optimization improves transactions-per-second and write amplification for workloads with many small batches.
The smaller the batch, the greater the (relative) improvement.

For example, with 256 transfers/batch the total bytes written is cut to 1/12th of the original, and transactions/second increases by 4.2x.